+++ lib/mdex/native.ex
@@ -53,6 +53,9 @@ defmodule MDEx.Native do
     mode: mode,
     force_build: System.get_env("MDEX_BUILD") in ["1", "true"]
 
-  def to_html(_md), do: :erlang.nif_error(:nif_not_loaded)
-  def to_html_with_options(_md, _options), do: :erlang.nif_error(:nif_not_loaded)
+  def markdown_to_html(_md), do: :erlang.nif_error(:nif_not_loaded)
+  def markdown_to_html_with_options(_md, _options), do: :erlang.nif_error(:nif_not_loaded)
+  def parse_document(_md, _options), do: :erlang.nif_error(:nif_not_loaded)
+  def ast_to_html(_ast), do: :erlang.nif_error(:nif_not_loaded)
+  def ast_to_html_with_options(_ast, _options), do: :erlang.nif_error(:nif_not_loaded)
 end