+++ native/comrak_nif/src/types/options.rs
@@ -1,4 +1,4 @@
-use comrak::ListStyleType;
+use comrak::{ExtensionOptions, ListStyleType, ParseOptions, RenderOptions};
 
 #[derive(Debug, NifStruct)]
 #[module = "MDEx.Types.ExtensionOptions"]
@@ -84,4 +84,68 @@ pub struct ExOptions {
     pub parse: ExParseOptions,
     pub render: ExRenderOptions,
     pub features: ExFeaturesOptions,
+}
+
+pub fn extension_options_from_ex_options(options: &ExOptions) -> ExtensionOptions {
+    let mut extension_options = ExtensionOptions::default();
+
+    extension_options.strikethrough = options.extension.strikethrough;
+    extension_options.tagfilter = options.extension.tagfilter;
+    extension_options.table = options.extension.table;
+    extension_options.autolink = options.extension.autolink;
+    extension_options.tasklist = options.extension.tasklist;
+    extension_options.superscript = options.extension.superscript;
+    extension_options
+        .header_ids
+        .clone_from(&options.extension.header_ids);
+    extension_options.footnotes = options.extension.footnotes;
+    extension_options.description_lists = options.extension.description_lists;
+    extension_options
+        .front_matter_delimiter
+        .clone_from(&options.extension.front_matter_delimiter);
+    extension_options.multiline_block_quotes = options.extension.multiline_block_quotes;
+    extension_options.math_dollars = options.extension.math_dollars;
+    extension_options.math_code = options.extension.math_code;
+    extension_options.shortcodes = options.extension.shortcodes;
+    extension_options.wikilinks_title_after_pipe = options.extension.wikilinks_title_after_pipe;
+    extension_options.wikilinks_title_before_pipe = options.extension.wikilinks_title_before_pipe;
+    extension_options.underline = options.extension.underline;
+    extension_options.spoiler = options.extension.spoiler;
+    extension_options.greentext = options.extension.greentext;
+
+    extension_options
+}
+
+pub fn parse_options_from_ex_options(options: &ExOptions) -> ParseOptions {
+    let mut parse_options = ParseOptions::default();
+
+    parse_options.smart = options.parse.smart;
+    parse_options
+        .default_info_string
+        .clone_from(&options.parse.default_info_string);
+    parse_options.relaxed_tasklist_matching = options.parse.relaxed_tasklist_matching;
+    parse_options.relaxed_autolinks = options.parse.relaxed_autolinks;
+
+    parse_options
+}
+
+pub fn render_options_from_ex_options(options: &ExOptions) -> RenderOptions {
+    let mut render_options = RenderOptions::default();
+
+    render_options.hardbreaks = options.render.hardbreaks;
+    render_options.github_pre_lang = options.render.github_pre_lang;
+    render_options.full_info_string = options.render.full_info_string;
+    render_options.width = options.render.width;
+    render_options.unsafe_ = options.render.unsafe_;
+    render_options.escape = options.render.escape;
+    render_options.list_style = ListStyleType::from(options.render.list_style.clone());
+    render_options.sourcepos = options.render.sourcepos;
+    render_options.experimental_inline_sourcepos = options.render.experimental_inline_sourcepos;
+    render_options.escaped_char_spans = options.render.escaped_char_spans;
+    render_options.ignore_setext = options.render.ignore_setext;
+    render_options.ignore_empty_links = options.render.ignore_empty_links;
+    render_options.gfm_quirks = options.render.gfm_quirks;
+    render_options.prefer_fenced = options.render.prefer_fenced;
+
+    render_options
 }